+++ pkg/deployments/src/network.ts
@@ -5,11 +5,11 @@ import { Network } from './types';
 
 const DEPLOYMENT_TXS_DIRECTORY = path.resolve(__dirname, '../deployment-txs');
 
-export async function saveContractDeploymentTransactionHash(
+export function saveContractDeploymentTransactionHash(
   deployedAddress: string,
   deploymentTransactionHash: string,
   network: Network
-): Promise<void> {
+): void {
   if (network === 'hardhat') return;
 
   const filePath = path.join(DEPLOYMENT_TXS_DIRECTORY, `${network}.json`);
@@ -24,7 +24,7 @@ export async function saveContractDeploymentTransactionHash(
   fs.writeFileSync(filePath, JSON.stringify(newFileContents, null, 2));
 }
 
-export async function getContractDeploymentTransactionHash(deployedAddress: string, network: Network): Promise<string> {
+export function getContractDeploymentTransactionHash(deployedAddress: string, network: Network): string {
   const filePath = path.join(DEPLOYMENT_TXS_DIRECTORY, `${network}.json`);
   const fileExists = fs.existsSync(filePath) && fs.statSync(filePath).isFile();
   if (!fileExists) {